provisioning1/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
docker==7.0.0
httpx==0.26.0
apscheduler==3.10.4
prometheus-client==0.19.0
python-json-logger==2.0.7
click==8.1.7
rich==13.7.0
pyyaml==6.0.1
python-dotenv==1.0.0

